Revert to 578ae8a6cf7

+++ src/attributes/ArrowPlottingAttributes.h
@@ -58,7 +58,7 @@ public:
 	double ratio_;
 	double max_speed_;
 	double min_speed_;
-	double thickness_;
+	int thickness_;
 	string unit_system_;
 	double unit_velocity_;
 	string legend_unit_;
